# app/docs/api-docs.yaml
openapi: 3.0.0
info:
  description: An offical Marine Management Organisation Developer API for Checking the validity of an Export Certificate
  version: 1.0.0
  title: MMO Check an Export Certificate
  contact:
    name: MMO Developer Support
    url: 'https://marinemanagement.org.uk/support'
    email: support@marinemanagement.org.uk
  license:
    name: OGL v3.0
    url: 'http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/'
servers:
  - url: '/'
    description: MMO Check an Export Certificate Service
tags:
  - name: Validation
    description: APIs for validating Export Certificate Details
  - name: Register Management
    description: APIs for managing the known Export Certificate Details
paths:
  /certificates/{certNumber}:
    get:
      tags:
        - Validation
      summary: 'Validate an Export Certificate by providing an Certificate Number'
      operationId: validateCertificate
      parameters:
        - name: certNumber
          in: path
          description: The Certificate Number to validate
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportCertificateValidity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
  /api/certificates:
    get:
      tags:
        - Register Management
      summary: 'List all Export Certificates registered within the system'
      operationId: listCertificate
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Successful
          content:
            application/json:
              type: array
              items:
                schema:
                  $ref: '#/components/schemas/ExportCertificateDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
  /api/certificates/{certNumber}:
    put:
      tags:
        - Register Management
      summary: 'Include details of an Export Certificate'
      operationId: includeCertificateDetails
      parameters:
        - name: certNumber
          in: path
          description: The Certificate Number to include
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      requestBody:
        $ref: '#/components/requestBodies/IncludeExportCertificate'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
    delete:
      tags:
        - Register Management
      summary: 'Remove details of an Export Certificate'
      operationId: removeCertificateDetails
      parameters:
        - name: certNumber
          in: path
          description: The Certificate Number to include
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      deprecated: false
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    ExportCertificateDetails:
      title: ExportCertificateDetails
      description: Details of an Export Certificate to store
      type: object
      properties:
        certNumber:
          type: string
          description: An Export Certificate Number
          example: 'GBR-2018-CC-123A4BC56'
        timestamp:
          type: string
          format: date-time
          description: The Export Certificate issue date in ISO8601 format
          example: '2019-07-07T13:53:40.257Z'
        status:
          type: string
          enum:
              - DRAFT
              - COMPLETE
              - VOID
      required:
        - certNumber
        - timestamp
        - status
    ExportCertificateValidity:
      title: ExportCertificateValidity
      description: Describes the validity of an Export Certificate
      type: object
      properties:
        certNumber:
          type: string
          description: An Export Certificate Number
          example: 'GBR-2018-CC-123A4BC56'
        timestamp:
          type: string
          format: date-time
          description: The Export Certificate issue date in ISO8601 format
          example: '2019-07-07T13:53:40.257Z'
        status:
          type: string
          enum:
              - DRAFT
              - COMPLETE
              - VOID
          description: The Export Certificate status
          example: 'COMPLETE'
        isValid:
          type: boolean
          description: Whether the Certificate Number relates to a valid certificate
          example: true
      required:
        - isValid
    Error:
      title: Error
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        stack:
          type: string
      required:
        - code
        - message
  responses:
    Success:
      description: Invalid parameters
      content:
        application/json:
          schema:
            title: Success
            type: object
            properties:
              message:
                type: string
                example: SUCCESS
    BadRequest:
      description: Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Access token not authorised to perform operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: An error occurred while performing the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  requestBodies:
    IncludeExportCertificate:
      description: Details of a valid Export Certificate
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExportCertificateDetails'
      required: true
    RemoveExportCertificate:
      description: Details of a valid Export Certificate
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExportCertificateDetails'
      required: true